move FetchRegistrationEntries into util pkg
